_attic/api/requirements-heavy.txt
# Reqs that don't change often and are allowed to take a while to build and push
flask-sqlalchemy==2.1
flask==0.11.1
gunicorn==19.6.0
haversine==0.4.5
ipython==5.0.0
pp-ez==0.2.0
pprintpp==0.2.3
pytest==3.1.3
pyyaml==3.12
requests==2.11.1
simplejson==3.10.0
sqlalchemy==1.1.3
structlog==16.1.0
